[CandlesManager] fix typing issues

+++ octobot_trading/exchange_data/ohlcv/candles_manager.py
@@ -155,7 +155,7 @@ class CandlesManager(util.Initializable):
                 self.open_candles[self.open_candles_index] = new_candle_data[enums.PriceIndexes.IND_PRICE_OPEN.value]
                 self.high_candles[self.high_candles_index] = new_candle_data[enums.PriceIndexes.IND_PRICE_HIGH.value]
                 self.low_candles[self.low_candles_index] = new_candle_data[enums.PriceIndexes.IND_PRICE_LOW.value]
-                self.time_candles[self.time_candles_index] = new_candle_data[enums.PriceIndexes.IND_PRICE_TIME.value]
+                self.time_candles[self.time_candles_index] = float(new_candle_data[enums.PriceIndexes.IND_PRICE_TIME.value])
                 self.volume_candles[self.volume_candles_index] = new_candle_data[enums.PriceIndexes.IND_PRICE_VOL.value]
                 self._inc_candle_index()
             except IndexError as e:
@@ -199,10 +199,10 @@ class CandlesManager(util.Initializable):
         max_handled_limit: int = self.max_candles_count if self.reached_max else max_limit
         if limit == -1:
             if max_limit == -1:
-                return np.array(data)
-            return np.array(data[:max_handled_limit])
+                return np.array(data, dtype=float)
+            return np.array(data[:max_handled_limit], dtype=float)
 
         if max_limit == -1:
-            return np.array(data[-min(limit, len(data)):])
+            return np.array(data[-min(limit, len(data)):], dtype=float)
         else:
-            return np.array(data[max(0, max_handled_limit - limit): max_handled_limit])
+            return np.array(data[max(0, max_handled_limit - limit): max_handled_limit], dtype=float)
